fix: moved modules folder to right place
