+++ .gitignore
@@ -10,3 +10,4 @@
 coverage
 *.log
 *.env
+*.tsv
