+++ src/main.rest.ts
@@ -5,9 +5,9 @@ import {
   createRestApplicationContainer,
 } from './rest/index.js';
 import { Component } from './shared/types/index.js';
-import { createUserContainer } from './shared/libs/modules/user/index.js';
-import { createCoordinatesContainer } from './shared/libs/modules/coordinates/index.js';
-import { createOfferContainer } from './shared/libs/modules/offer/index.js';
+import { createUserContainer } from './shared/modules/user/index.js';
+import { createCoordinatesContainer } from './shared/modules/coordinates/index.js';
+import { createOfferContainer } from './shared/modules/offer/index.js';
 
 async function bootstrap() {
   const diContainer = Container.merge(
