+++ src/shared/modules/coordinates/coordinates-service.interface.ts
