+++ src/shared/modules/coordinates/coordinates.entity.ts
