+++ src/shared/modules/offer/dto/create-offer.dto.ts
