+++ src/shared/modules/offer/offer.container.ts
@@ -1,6 +1,6 @@
 import { Container } from 'inversify';
 import { OfferService } from './offer-service.interface.js';
-import { Component } from '../../../types/index.js';
+import { Component } from '../../types/index.js';
 import { DefaultOfferService } from './default-offer.service.js';
 import { OfferEntity, OfferModel } from './offer.entity.js';
 import { types } from '@typegoose/typegoose';
