+++ src/shared/modules/user/default-user.service.ts
@@ -3,8 +3,8 @@ import { DocumentType, types } from '@typegoose/typegoose';
 import { UserEntity } from './user.entity.js';
 import { CreateUserDto } from './dto/create-user.dto.js';
 import { inject } from 'inversify';
-import { Component } from '../../../types/index.js';
-import { Logger } from '../../logger/index.js';
+import { Component } from '../../types/index.js';
+import { Logger } from '../../libs/logger/index.js';
 
 export class DefaultUserService implements UserService {
   constructor(
