+++ src/shared/modules/user/user-service.interface.ts
