+++ src/shared/modules/user/user.container.ts
@@ -1,6 +1,6 @@
 import { Container } from 'inversify';
 import { UserService } from './user-service.interface.js';
-import { Component } from '../../../types/index.js';
+import { Component } from '../../types/index.js';
 import { DefaultUserService } from './default-user.service.js';
 import { UserEntity, UserModel } from './user.entity.js';
 import { types } from '@typegoose/typegoose';
